Revise directory structure and module boundaries
The `api` directory contains data types and functions shared between the
explorer client and the server. They must be platform-agnostic and
serialisable.

The `explorer` directory contains Preact components and hooks and
client-side helper functions.

The `server` directory contains the Vitest integration and adapters and
the request handlers for the Polka web server.

`ProjectEvent` has been renamed to `Event` to align it with
`EventStream` and `EventLogger`, undoing the change in commit 7680f9f.